+++ service/service/DEPRECATED_service_test.go
@@ -39,7 +39,7 @@ var _ = Describe("DEPRECATEDService", func() {
 			Expect(server).ToNot(BeNil())
 			server.AppendHandlers(
 				CombineHandlers(
-					VerifyRequest("POST", "/auth/serverlogin"),
+					VerifyRequest("POST", "/serverlogin"),
 					VerifyHeaderKV("X-Tidepool-Server-Name", *provider.NameOutput),
 					VerifyHeaderKV("X-Tidepool-Server-Secret", serverSecret),
 					VerifyBody(nil),
@@ -50,7 +50,8 @@ var _ = Describe("DEPRECATEDService", func() {
 				"address":             server.URL(),
 				"server_token_secret": authTest.NewServiceSecret(),
 				"external": map[string]interface{}{
-					"address":                     server.URL(),
+					"authentication_address":      server.URL(),
+					"authorization_address":       server.URL(),
 					"server_session_token_secret": serverSecret,
 				},
 			}