+++ task/service/service/service_test.go
@@ -45,7 +45,7 @@ var _ = Describe("Service", func() {
 			server = NewServer()
 			server.AppendHandlers(
 				CombineHandlers(
-					VerifyRequest("POST", "/auth/serverlogin"),
+					VerifyRequest("POST", "/serverlogin"),
 					VerifyHeaderKV("X-Tidepool-Server-Name", *provider.NameOutput),
 					VerifyHeaderKV("X-Tidepool-Server-Secret", serverSecret),
 					VerifyBody(nil),
@@ -56,7 +56,8 @@ var _ = Describe("Service", func() {
 				"address":             server.URL(),
 				"server_token_secret": authTest.NewServiceSecret(),
 				"external": map[string]interface{}{
-					"address":                     server.URL(),
+					"authentication_address":      server.URL(),
+					"authorization_address":       server.URL(),
 					"server_session_token_secret": serverSecret,
 				},
 			}